+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 ----
 * Optional `--url` parameter in `install:run` command if url is set in configuration file.
 * Add `--email` option for `user:update` command to set email address of a user.
+* Add `--ask-password` option for `user:update` command to ask for password
 
 1.9.1
 ----